+++ vitest.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   test: {
     globals: false,
     environment: 'node',
-    include: ['__tests__/**/*.test.ts'],
+    include: ['__tests__/**/*.test.ts', 'packages/*/__tests__/**/*.test.ts'],
   },
 })